+++ CHANGELOG.md
@@ -8,6 +8,7 @@ customers cannot upgrade their bootloader, its changes are recorded separately.
 
 ### [Unreleased]
 - Bitcoin: add support for payment requests
+- Bitcoin: allow multisig accounts at arbitrary keypaths
 - Ethereum: allow signing EIP-712 messages containing multi-line strings
 
 ### 9.18.0